pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "care-orchestrator"
version = "0.7.0"
description = "AI-Native Vertical Engine for Reducing U.S. Healthcare Administrative Friction"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Bahtiyar Aytac" },
]
keywords = ["healthcare", "ai", "hipaa", "fhir", "mcp", "anthropic", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "fhir.resources>=7.1.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]
server = [
    "uvicorn[standard]>=0.32.0",
]

[project.urls]
Homepage = "https://github.com/bahtoo/claude-care-orchestrator"
Repository = "https://github.com/bahtoo/claude-care-orchestrator"
Issues = "https://github.com/bahtoo/claude-care-orchestrator/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/care_orchestrator"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN001", "ANN201", "ANN202", "S106", "S105"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff.lint.isort]
known-first-party = ["care_orchestrator"]
